Move classes for tests to test package.
