+++ src/test/java/oripa/domain/fold/subface/SplitFacesToSubFacesConverterTest.java
@@ -27,6 +27,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import oripa.domain.fold.subface.test.OriFaceFactoryForTest;
+
 /**
  * @author OUCHI Koji
  *
